requirements.txt
gradio>=4.44.0
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
azure-identity>=1.17.0
azure-search-documents>=11.6.0
openai>=1.40.0
